# swagger/swagger.json
{"swagger":"2.0","info":{"title":"The customer service of Westeros Shop","description":"The customer service of Westeros Shop","version":""},"host":"localhost:6543","schemes":["http"],"consumes":["application/json","application/xml","application/gob","application/x-gob"],"produces":["application/json","application/xml","application/gob","application/x-gob"],"paths":{"/customers/":{"get":{"tags":["customer"],"summary":"list customer","description":"Lists all customers","operationId":"customer#list","produces":["application/vnd.goa.example.customer+json"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/GoaExampleCustomer"}},"404":{"description":"Not Found"}},"schemes":["http"]}},"/customers/{customerID}":{"get":{"tags":["customer"],"summary":"show customer","description":"Get customer by id","operationId":"customer#show","produces":["application/vnd.goa.example.customer+json"],"parameters":[{"name":"customerID","in":"path","description":"customer ID","required":true,"type":"integer"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/GoaExampleCustomer"}},"404":{"description":"Not Found"}},"schemes":["http"]}},"/swagger/{filepath}":{"get":{"summary":"Download public/swagger/","operationId":"swagger#/swagger/*filepath","parameters":[{"name":"filepath","in":"path","description":"Relative file path","required":true,"type":"string"}],"responses":{"200":{"description":"File downloaded","schema":{"type":"file"}},"404":{"description":"File not found","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}}},"definitions":{"GoaExampleCustomer":{"title":"Mediatype identifier: application/vnd.goa.example.customer+json; view=default","type":"object","properties":{"id":{"type":"integer","description":"Unique customer ID","example":5804055118684468889,"format":"int64"},"name":{"type":"string","description":"Name of wine","example":"Qui praesentium est omnis praesentium corrupti."}},"description":"A customer of Westeros Shop (default view)","example":{"id":5804055118684468889,"name":"Qui praesentium est omnis praesentium corrupti."},"required":["id","name"]},"error":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"code":{"type":"string","description":"an application-specific error code, expressed as a string value.","example":"invalid_value"},"detail":{"type":"string","description":"a human-readable explanation specific to this occurrence of the problem.","example":"Value of ID must be an integer"},"id":{"type":"string","description":"a unique identifier for this particular occurrence of the problem.","example":"3F1FKVRR"},"meta":{"type":"object","description":"a meta object containing non-standard meta-information about the error.","example":{"timestamp":1458609066},"additionalProperties":true},"status":{"type":"string","description":"the HTTP status code applicable to this problem, expressed as a string value.","example":"400"}},"description":"Error response media type (default view)","example":{"code":"invalid_value","detail":"Value of ID must be an integer","id":"3F1FKVRR","meta":{"timestamp":1458609066},"status":"400"}}},"responses":{"NotFound":{"description":"Not Found"},"OK":{"description":"OK","schema":{"$ref":"#/definitions/GoaExampleCustomer"}}}}